[REFACTOR][I] Move IntelliJ event handlers to a separate package
Moves all event handlers dealing with IntelliJ events to the package
eventhandlers. The package contains sub-package which represent the
differentiation between the different resources the events deal with.

Adds package-info.java files to the added packages describing their
purpose.

Subsequently changes the visibility of used methods from package-private
to public. As these methods should still only be used internally, a note
was added to the javadoc of the changed methods.

Adjusts the visibility of the moved handlers as necessary.

Adjusts the test package-structure to match the new eventhandler package
structure.