[INTERNAL][I] Unify access to en/disabling handlers in EditorManager
Creates methods to enable/disable all handlers for a specific resource.
Replaces direct access to the setEnabled(boolean) method of the handlers
with calls to the new wrapper methods.

This is done to make it easier to keep the state of all handlers for a
resource in sync when adding a new one. Otherwise, we would have to
manually search where the other handlers for the same resources are
enabled/disabled.

Moves the existing methods for this purpose to the same position as the
rest of the methods.